Rename and document a #define for downstream use
Downstream packaging targeting reproducible builds may set file mtimes
to a stable value, e.g. 0; see #1702. Unfortunately .des cache
invalidation relies on mtime, so this ends up breaking under these build
conditions. This commit is aimed to refine and simplify the advice I
gave in that issue thread: versioned des cache directories are now our
(or at least my) official recommendation for dealing with this
situation. This change makes it clearer that this option is not tied to
DGL, and provides a hint about the option directly in the Makefile. I've
tested this and it seems to work fine on its own.

Resolves #1702