Simplify flight
There are now very few reasons for the player to micromanage flying
status. This commit therefore removes the interface for doing it. Like
monsters, if the player *can* fly then they are flying:

* flying items cause you to fly while wearing them, rather than
  requiring an evocation.
* Species and form flight -> always flying, like Djinni.
* Emergency flight and potions should be unchanged.
* the abilities to start and stop flying are removed, as is
  ATTR_FLIGHT_UNCANCELLABLE and ATTR_LAST_FLIGHT_STATUS.

The one remaining special case I know of (pointed out by Kate) is that
Op and Mf have incentives not to fly; I'll deal with that in a separate
commit.

The old flying code had so many weird interactions that there may still
be special cases that don't work right here...