Rework scale mutations
There were a huge number of scale mutations, and characters were limited to a
maximum of 3 levels across all of them. This was intended to prevent
heavily-mutated characters from gaining massive AC bonuses from lots of types
of scales, but as a result it was unlikely to reach the third level of any
individual scale mutation (which is where many of them were differentiated by
providing a resistance).

Instead remove the scale cap of 3 levels, but make most of the scale variations
that provide AC and a resistance unique to Demonspawn - distortion field and
the rF/rC/rPois/rElec/rCorr/+HP scales are now all DS-only.

Iridescent scales (+AC only) are removed from the DS mutation pool and remain
as a random mutation for other species. Bone plates, thin skeletal structure
and sturdy frame also remain available as random mutations, and no longer
conflict with each other.