Allow Ru's -Scroll as a random mutation replacing Blurry Vision
Blurry Vision is a strong negative mutation but is quite messy in how its
effect works. Making an action that isn't normally delayed into a delay has
strange consequences, with the timing of when the world and monsters get to
react compared to when the scroll comes into effect changing in a confusing
way.

Instead allow Ru's existing sacrifice scrolls mutation as a random bad
mutation, since this has a similar strong effect of limiting scroll usage but
is much clearer in its effects.