Remove the blink mutation
It was the final remaining random mutation that provided an activated ability,
most of which were removed in the distant past due to providing fairly
forgettable effects. The ability to blink is strong but still duplicates both
a spell and randart property, so isn't really distinct enough to be worthwhile
as an exception to the rule of mutations being passive.