Slow down Sif book gifts
Sif gifting occured very frequently at 5* piety (relative to the gift
timeout). As a consequence, reaching 6* piety for Sif took inordinately
long (especially if the player was also spending piety). For mechanical
purposes, the practical impact was a brake on how many times Exegises
could be used, but if we want to control that we should use piety cost.

From the gamefeel side the consequence was awful. Players like reaching
6* with their god, it feels good.

This halves the chance of getting a gift from Sif when the gift_timeout
is 0, which brings her gifting rate closer to Trog and Oka.