Revert "Prevent draconians from getting armour troves"
This commit didn't actually properly do the thing it claimed to do
(trove_armour_1 could generate for draconians), and it also wasn't really
necessary since draconians would still generally get lots of good aux armour
from these troves.

It also tried to prevent draconians from being asked for enchanted dragon
armour as a trove toll, but this only worked on <XL7 draconians with no colour
yet. The dragon armour toll functions mainly as a tax on enchant armour
scrolls, which are just as useful for draconians as for a number of other
species with reduced armour slots, so is fine to keep enabled.

This reverts commit 62faa89a7ed0f09c553b4d46906d7504b5d1229d.