Remove decorative altars from early D vaults
This either adds conditional placement, simply deletes the altar, moves
the depth range out of D, or turns the vault into a proper overflow
temple depending on the vault tweaked.

This is part of a sequence of commits giving the dungeon builder more
control over the placement of early altars where they have a large strategic
impact.