Shift overflow temples deeper, increase ecumenical altars
Ecumenical altars suffer from the fact that they don't always show up
before identified altars; the choice becomes more about a conduct than
the chance to gamble for a serious advantage. This commit moves all
overflow altars to D:3-. Combined with the previous commits, this means
that the only way for an altar to appear on D:1-2 is via the basic_altar
vault, which is given an explicit chance of 33% on D:2.

Ecumenical altars are given a 50% chance for each of D:1-3. This makes
for an 87.5% chance of seeing one, and a 50% chance of seeing one before
any identified altar.

There is some concern that placing an ecumenical altar on D:1 will
encourage startscumming, but a player who would quit on D:1 because they
didn't like their ecumenical roll would probably also quit on D:2.