Tweak the fsim test
This test was using a mifi and is therefore impacted by shield skill in
a way that is somewhat orthogonal to the point of the test (e.g. the
results changed with 0b64aa269c0). Switch to a migl.

(cherry picked from commit a1c911b16d80c32dc23f4a437d50b2bc023b383e)