Make WJC Heavenly Storm grant bonus EV
For such an expensive active ability, it was fairly difficult to use
effectively, since it tries to encourage staying in combat against multiple
foes for an extended time (while making a lot of noise) to gain its
benefits. Grant it a defensive bonus alongside the offensive one to
encourage that use, and slightly reduce the noise to the level of normal
shouts.