New ecumenical altar vaults
We still need more of these, but this is a start. A collection of small
vaults placing the faded altar plus a bit of decor, some of which
suggests an old temple (webs, ruins, etc.) or place of worship.

There still aren't enough, so to keep them fresh the weight of the basic
vault is increased. Hopefully by the next release we'll have plenty of
vaults.