Disable ood fuzzing for 8 and 9 spawns.
8 and 9 are specific requests by vaults for an out-of-depth monster with
a certain depth modification. Fuzzing can, early on, turn a 9 into an 8.
This was previously less noticable in early D because of the non-monster
weight in the spawn table. With that removed a D:4 9 can get fuzzed into
something worse than an 8 on a more frequent basis.

0 is not adjusted since it is supposed to be a "typical" spawn, so
should behave just as if the dungeon builder is placing the monster.