Simplify the foxfire, And replace foxfire card with swarms card
This card had a complicated function. There are many different types of monsters that appear on this card, but there was no reason to do so.  And some of the summoned monsters appeared hostile. The name was also not intuitive, making it difficult for users to understand what the card was doing. The user had too many considerations to make this card useful.

So I'm going to simply trim this card.
The basic feature of this card is that it summons many monsters. Therefore, I want to emphasize this characteristic clearly. Now this card summons a very large swarm of bees. In addition, I will change the name to swarms to make it easier to understand what this card is summoning.

If this change is not enough for cards to be useful, I think we can mix one or two queen bees.