Replace wands of random effects with Xom's chesspieces
This commit moves random effects to an XP evoker, and makes it
non-targeted. This is partly for flavour (in order to re-use the amazing
chesspiece tiles), and partly for mechanical distinction. Random
Effects' effect is not something that would be good to teach monsters
(turning random goblins into possible malmutators/paralysis checks
etc.); and random effects was the only remaining wand that monsters
didn't know how to use.

This gives a clear distinction between misc evockers (monsters don't
know how to use) and wands (monsters do know how to use).

The change is also an experiment in high-charge misc evokers, to get a
broader feel for how they play. This is a warm-up for possibly
converting all wands to xp-charging.