New consumables panel for Web Tiles
The main goal of adding this panel is to give the player a quick and
easy way to see all the items that can help them succeed in a fight.

It would be ideal to have the full inventory panel, similar to Local
Tiles, but there are a couple of problems. The first is a lack of screen
space: unless you have a big monitor, the choice is either the monster
list (Web Tiles) or the inventory panel (Local Tiles). The second
problem is that only about a half of inventory items are useful in
fights. A filtered view of the inventory will take less space and
present only combat-relevant items.

This commit adds a new panel, which can be configured via four options:
  * `consumables_panel`: types of items to show on the panel;
  * `consumables_panel_filter`: a list of regular expressions for
    filtering items out;
  * `show_unidentified_consumables`: whether to show unidentified
    scrolls and potions;
  * `consumables_panel_scale`: allows to change the size of icons on
    the panel.

The panel is displayed in the top left corner, partially covering
the top row of tiles on the map. Non-Barachim chars should always have
at least one out-of-sight row at the top. For Barachim, the game can
show 18 rows of tiles with one out-of-sight row at the top as long
as you have 736 pixels vertically.