Filter consumables on the crawl binary side (advil)
This allows to use the same regex parser that the crawl binary uses,
which should be faster. Also, it makes the code more portable to
other build targets.

Now `TilesFramework::_send_item()` does the filtering and sends
an extra flag with each item, `qty_field`. If it's either "plus" or
"quantity", the web client(s) will show such items on the panel.