Hilbert curve vaults for Spider by nicolae (12539)
Moved these to their own file, since it's a large series of vaults and
subvaults. I removed the placement down-weighting, since it's not an
extreme vault in any way, and it has a uniq tag to prevent it from
placing more than once.

These vaults can create large rectilinear areas when the layout embeds
them in large areas of rock. To some extend this might be helped by
adding a "fuzzed" CLEAR border to the map, but that would apparently
introduce potential closet areas, and it wouldn't fully solve the issue.
Seems to be a problem all larger vaults have with the cave-like layouts
especially, and I'm not sure if there's an easy solution to that.