Reduce Makhleb's heal-on-kill
Instead of healing for (roughly) the monster's HD * 1.5, heal for (roughly) invocation skill + HD * 0.5.

This is intended to help with the highly visible runaway Dj/DD games, and also with the less-talked-about problem where Makh is a really strong choice for almost any character.  It also incentivizes training Invocations, which could help makh's active abilities become less of a sideshow.