Add an option to reposition the consumables panel (Implojin)
To make the panel less intrusive for more screen resolutions and sizes,
add the `consumables_panel_orientation` option.

When set to 'horizontal' (the default), the consumables panel will be
placed along the top edge of the screen. When set to 'vertical',
the panel will be placed along the left edge.