Update dispellable monster enchantments
Makes lots of monster enchantments dispellable that weren't previously.
Some could be dispelled on the player but not on monsters, and some were
monster-only magical effects that didn't have any particular reason to not
be dispellable. Also makes some effects non-dispellable: the frenzy effect
from darts (for consistency with other darts), and silver corona, pain bond
and idealise (for consistency with other player god abilities).

In a rare case of favouring the player over monsters, this also makes
monster priestly buffs generally be dispellable - on the player,
god-granted buffs can't be dispelled, but on monsters this isn't feasible
to do consistently, since spells can be either wizardly or priestly in
origin. The previous implementation was very arbitrary (for example, the
regeneration portion of Trog's Hand could be dispelled but not the
willpower; Donald's priestly might and haste could be dispelled, but deep
elf high priests' brilliance aura couldn't be).