Fix line_reader wide char wrapping (dilly)
If the line_reader wrapping code hit a wide character on the wrap
column, the wrapping code would go into an infinite loop, and the
process would need to be killed. This could be triggered in search as
well when the previous search was the exact right length to put the
cursor for a following search (since it repeats the previous search) at
the wrap column, and the first character of the new search was a wide
character, then the bug would trigger. Because webtiles still uses the
line_reader console ui behind the scenes, the bug was triggerable in
webtiles this way by just having strings of the right length. This
probably didn't come up for normal usage, but apparently people use the
search box to chat to spectators. (Even worse, a webtiles game killed by
the server in this state doesn't save, so the bug can be used to roll
back a game on demand.)

This commit fixes the infinite loop bug in a pretty straightforward way
(and also tries to better guard against infinite loops there). It also
fixes some fiddly issues with cursor positioning in the case where the
last character of a line is a wide char.