Let the arena create an ironbound beastmasters on request.
Previously, the game terminated (with game_ended_with_error()) if the user attempted to create an arena with a M_REQUIRE_BAND monster which didn't create a band in the same dgn_place_monster() call.

This change means that the monster is created as if the M_REQUIRE_BAND flag wasn't there.