Make the Wrath of Trog worthwhile
Wrath of Trog is very thematic with ANGRY: 50, but suicidally dangerous if you cannot manage the *Rage.  This is 5.55x more frequent than Berserkitis 3.  Other unrands with *Rage, besides have ANGRY of only 5, also have beneficial properties such as super-vampiricism, Slay+6, and berserking allies.  Wrath of Trog, though, is simply a +8 antimagic battleaxe.  This hardly compensates for berserking the wielder 10x as often.

This PR brings Wrath of Trog's ANGRY down to 5 and adds beneficial properties to it, to match the other *Rage unrands.  In keeping of the theme with this being Trog's favorite weapon, instead of giving it Troggish properties such as Regen+, Will++, or +Rage--things Trog wouldn't need himself from a weapon--just make it very powerful weapon with no properties other than the *Rage.  Trog's favorite weapon ought, after all, to be as good or better as the best gifts he used to give his followers!

This is accomplished by adding BASE_DAM: +3 and BASE_DELAY: -4 to its entry in art-data.txt, giving it the base damage of an executioner's axe with the base delay of a dire flail.  The enchantment is not changed.  Attached are five graphs comparing the new axe at skills 14, 20 and 26 to a +9 battleaxe of speed and to a +9 anti-magic executioner's axe, plus the fsim.csv.  The higher the target AC and EV, the better Wrath of Trog's relative performance.

The fsim settings were fsim_scale = fighting and fsim_rounds = 10000.  Wrath of Trog was changed in atr-data.txt to comment out the ANGRY line for fsim testing and to add BASE_DELAY = -4 and BASE_DAM = +3, then recompiling just in case.  Player stats were Strength 30 and Dex 20, wearing plate mail with Armour skill 20.  Target monsters were fungus (AC/EV of 0/0), tentacled monstrosity (5/5), wraith (10/10), orb guardian (13/13) and orb of fire (20/20).