Implement petrified trees
Instead of a vault recolour, let's make petrified trees a first class
feature. The name implies that they should behave like rock in some ways
(be valid LRD targets, be diggable). Implementing them as vault re-tiles
and re-names of statues does this, at the expense of clarity for console
players. The expense is even higher for screen-reader players, who might
go looking for 't' with a reading cursor only to find '8' instead.

As a first class feature, petrified trees can behave like their woody
bretheren in some ways, be vulnerable to LRD instead of fire, and
interact properly with our feature naming/colouring/glyphing systems.

On the vault decorating side, tree_dead is now the variant tile for the brown
colour of trees, and no living tree is assigned darkgrey in console.

There are also vague future plans that might use such a feature: i-x's
Elven Crypt sketch or expanding Desolation into a (rune?) branch. This
feature is implemented in the spirit of "if you build it they will
come".