Trade Ice Beast for Metabolic Englaciation in the IE Start (Nomi)
Summons have a way of defining a book, especially the mid-level options
like Summon Ice Beast. Lighting Spire was briefly considered and then
removed from the Air Elementalist start for just this reason.

While the positional-magic changes to the book of Frost have made the
book a bit less "the book of SIB" the background still very much ends
up leaning on the spell.

Metabolic Englaciation is an interesting spell and a good fit for the
starting background. AOE slow interacts well with the other starting
spells, but the price is sufficiently high and there's no direct damage,
so it shouldn't outshine in the way SIB does.