Support for mac universal builds targeting apple ARM
It turns out that apple clang supports cross-compiling in an extremely
straightforward way for arm64 vs x86_64; building a universal binary
amounts to cross-compiling and then merging binaries with the `lipo`
utility. This also means that producing a universal build on an
up-to-date x86 device should also be possible (perhaps even in CI?).
However, one caveat is that this forces our back-compatibility hand a
bit -- I'm not exactly sure how much since I don't have the devices to
test, but the choice of 10.12 is based on apple docs. I haven't removed
the non-universal app building scheme so we would have the option of
distributing multiple builds.

Because of the somewhat baroque Makefile setup that exists for mac
builds (possibly in combination with my limited knowledge of what the
best way to accomplish this build might be) this is a relatively ugly
patch. I've also found rebuilds to be a bit flaky for reasons I don't
fully grasp, e.g. sometimes a full clean or clean-contrib is needed when
changing targets.

Some other minor cleanups, removing files that (as far as I can tell)
are no longer used.