Replace hydra form with air walk.
Hydra form was intended to be a heavily power-scaling form that made up for its
bad defenses with cleaving melee and heal on kill. In practice, it was hard to
obtain enough power for hydra form to outdamage to other forms, let alone find
situations in which its benefits made up for poor ac and awful evasion.

Air walk will attempt to fill the same niche as a lategame form that rewards
high spellpower, but hopefully do it a bit better. Air walk melds non-jewelery
slots, provides 2 + power / 5 base damage elec branded cleaving melee (breakeven
pre-brand damage with blade hands at 95 power and with dragon form at 140),
10 + power / 10 AC, retains player size and adds a power / 10 ev bonus, provides
flight and relec, and has an activatable "blinkbolt" ability that deals elec
damage and blinks the player to the end of its path, but must target a monster
and has a cooldown.

The new spell is level 6 transmutation/air, which makes spellpower more expensive
in terms of exp but does allow easier enhancement. It is probably still very
niche, but should be better in its niche. Numbers are open to adjustment.
Additionally, it needs cool tiles for the form and ability.