Tweak randart property availability
Attempts to simplify which properties can appear on which types of
equipment. Some properties which shouldn't appear on quick-swappable items
can now generate on amulets as well as armour (in particular Rampage, Harm,
Regen, -Cast and -Tele). rCorr can now appear on any item that doesn't have
*Corrode. +Invis can now only appear on armour and amulets, to limit
swapping to activate it. Some handling for *Tele is removed since it
doesn't generate randomly.