Rework monsters in a lemuel castle vault
The vault lemuel_castle_with_subvaults has monster placement that's not
much been updated, and still has logic based on Dungeon having 27
levels. This commit reworks the monster and loot placement, giving it
better scaling with depth, with appropriately themed monsters at each
location. It also cleans up the tagging and supporting lua code,
organizing it similarly to our current DES standards.

For monsters in Dungeon, we use the same scaling as ghost vaults, since
this vault places over the same range. Previously it had a specific
theme for each depth (e.g. gnolls then later orcs), but something so
specific is more difficult to scale. Instead we use a mix of gnolls,
orcs, kobolds, and other Bailey-loving humanoid monsters to give us good
variety of monster classes. We introduce different types "guard dogs" as
the depth changes, using hounds after jackals, and also adding in some
howler monkeys, since those are great at rousing the castle army. In the
late dungeon hell hounds and even raiju make an appearance at the back
of the castle.

For Depths, we keep the "giant" theme, since there are enough such
monsters to choose from, and use mostly dragons as the guardian beasts
instead of hound genus monsters, however one may found a shrike or X at
the back of the castle moat. On Depths:5 it places the Zot entrance and
draconian and dragon monsters as before, but with some tweaks to the
composition of those. Keep the death drakes, but use real dragons
instead of a bunch of low-level drakes. Don't thin out the number of "1"
monsters drastically only for Depths:5, rather removing about 8 of these
from the original lemuel subvault, since that one places so many more
than the other subvaults. It still places the most monsters (and the
most loot), just not so many more compared to the other subvaults.

For loot, scale up the type and quality with depth instead of the
thining out numbers earlier. We don't change the monster count  with
depths, so the overall loot count should likewise remain constant.