Try to fix homebrew for mac CI runner
Homebrew is migrating to github for storage, because their previous
provider is shutting off service May 1. There's a brownout right now "to
warn users about May 1st as sunset date for Bintray" (where the warning
takes the form to downstream users of a completely uninterpretable 502
error, but ok). This has revealed that the default mac runner uses a
version of homebrew that is not updated with the correct download
location.  There's various plausible ways to fix this but in absence of
any official guidance I'm starting with just inserting the official
homebrew setup action in hopes that this fixes things up.