Fix sacrifice validity issues + document
A number of crucial checks for sacrifice validity were being done in
`_random_valid_sacrifice`, which it turns out is not used for all
sacrifices (only for ones that involve multiple possible mutations).
This commit abstracts those checks into a static function and also
applies them to all the other sacrifice possibilities in
_sacrifice_is_possible directly. In addition, I've tried to add a few
commenting hints throughout to help this code be more easily understood.
Probably this could use a more thorough refactor.

I've also added another check to ensure that piety is always 0 for
useless skills; I don't think this should do anything on top of the
other checks but it does seem like a useful sanitizer.

Fixes mantis 12545 as well as a similar bug on felids. This happened
when c3fac4b18bf removed a sac armour-specific check into
_random_valid_sacrifice. I'm not sure if anything else would have been
affected, because some or all of the other checks maybe were only
relevant to health/essence/arcana mutations? But this will future-proof
the code a bit as well.