Simplify GDR
Guaranteed damage reduction is intended to smooth out some of the
nastiness in Crawl combat. It guarantees that your AC will never
roll below some % of its max value when defending against melee attacks
and mundane projectiles (not against spells, etc).

This is a mostly reasonable idea, but it's implemented bafflingly: it
uses your base AC OR a hodge-podge of special cases based on your species
and/or form, meaning that it's both spoilery and weird. Some AC is
effectively worth more than others. Not great!

Instead, simplify it to be based purely on the player's current AC.
(Specifically, AC^(1/4) * 16.)

Previous example values:

Robe:                 0%
Scale mail (ie Fi):  28%
Plate:               39%

New values:

AC 6 (ie Fi):        25%
AC 20:		     34%
AC 50:               43%

This is a buff to some things (anything that didn't have GDR before) and a
nerf to other things (early plate). It's all over the place. Chaos reigns.
(But in practice I suspect the difference will be mostly unnoticable.)