WIP: add mojo/mjolnir/mjollnir to the game
However you wanna spell this item, it's the so-called "hammer of thor".

The main idea is that it's a melee weapon that also can be thrown, and
it has the returning property.  Advice greatly welcomed as i work on
this; i'll force-push as needed to the PR until it's done.