Allow specifying GAME at the make command line
This is sometimes requested by downstream packagers in one way or
another, e.g. as in #464. That PR hardcodes a change for tiles, which
would be pretty reasonable if one were writing this from scratch. But,
there are many secondary build processes that assume the binary is named
`crawl`, so it's a bit tricky to make that change as requested without a
lot of followup pain. Instead, this allows e.g. packagers to invoke the
build process as: `make TILES=y GAME=crawl-tiles` to have a one-step
build process that produces a differently-named binary without touching
`crawl`.