Standardize app icon location/naming
These were a bit scattered, leading to potential confusion for
downstream packagers (and me!). I'm not sure if dat/tiles is really the
best place for this, because that will get copied into installs, but it
is at least a place where the most icons already existed. As a
convenience, this pulls in the (currently unused) 48x48 icon from #464.
The svg in that PR was identical to the one that was in util/ so no need
for an additional copy (possible that the copy in util was simply
missed, it was far from an obvious location). For downstream packagers
in the future: rather than adding duplicate icons to the repo, the
preferred process would be to copy them out of `dat/tiles` similar to
how the mac and windows builds now work, and to add any new icons to
this location -- though I think at this point we might have enough to
cover every base!

(Commit caveat: not tested yet on windows, I'm going to let CI do
that...)