fix(tiles): Skip pre-layout message rendering
This is basically a followup to 5a74e75d8f72.  Those crashes in local
tiles revealed that the message window is getting rendered before tiles
has a layout, so the message window is width 0; this is basically
harmless but triggers pathological linebreaking behavior that it's
better to skip (and has the potential to trigger rendering glitches,
though I don't think it will currently). Re the removed comment there: I
have verified that the current wrapcprint behavior should prevent
infinite loops even with width 0, as long as linebreaks are correctly
added; the loop will always abort if it runs out of y space. The bug in
cfdca905f5 involved failing to correctly add a linebreak. So restoring
the wrapcol ASSERT should be unnecessary.