docs: Document some commit conventions
There's a bunch of pitfalls that I see coming up in PRs again and again
to do with commit messages, and this tries to make them explicit. My aim
here is to document conventions that we do have, but that are often
unstated or only stated in chat / PR feedback -- but if anyone feels
that I got things wrong or too prescriptive here I'm happy to be
corrected!