Remove the stochastic torment resistance mutation
It's removed from the black mark facet and not slated to be added to any
of the demonspawn facet re-works. hellmonk suggested making this
mutation deterministic and give 50% resistance, to be
used in future demonspawn. This is already done by MUT_TORMENT_RESISTANCE
level 1 so there's no need to re-create the effect.

For save compatibility the old mutation applies the new effect.