fix!: reorder demonspawn mutations added in #1852
Commits in that PR added new mutations to the middle of the enum; this
has save compat consequences because this enum is serialized as an int.
This is a bit of an emergency fix to prevent further breakage; the
commit has already gone live on cko, and save compat will get even worse
for impacted games, but we really need to keep this from impacting other
servers.