fix: more general code for species mutation fixup
This will more consistently remove species mutations fixed up this way
if they shouldn't be there, as well as add them if they should. This
code is still not very general (in principle this should be readable off
of the species def, but it's tricky to get that right for all cases),
and doesn't handle ds muts.