Fix an error message when creating a malformed dancing/spectral weapon.
Previously, if you using the wizard mode create monster command with
"dancing weapon; stone" on a build with asserts compiled in, the game
crashed with "ASSERT(wpn) in 'dungeon.cc' at line 5159 failed.". The
same happens for any non-weapon, or even "nothing", and the same if a
vault was created which included such a monster.

With this change, it prints "Monster 'dancing weapon' requires a
weapon." instead in wizard mode, and the game cannot start if any vault
contains a monster like that.