Add a help option for the wizard mode create monster command.
Entering ? as the monster name causes some help text to appear. This is
derived from the MONS section of docs/develop/levels/syntax.txt, but
with a few changes.

1. Use colour - yellow for headings, white for keywords which can be
entered as part of the command, and cyan for other things such as text
the game produces in response to a command, or file references.

2. Remove quotation marks from around keywords. I haven't done this with
cyan text, as this contains longer phrases which may be harder to
separate.

3. Don't give advice on how the command should be used (such as "use
this sparingly").

4. Rearrange the paragraphs a bit, and add a new opening paragraph.

5. Add lists of valid characteristics for mutant beasts and ugly things.