Nerf Makhleb heal-on-kill (cebolla, kate-)
This passive is runaway strong and outshines Makhs shiny actives.  On
every kill that passes the piety threshold (random2(you.piety) >= 30)
(at high piety more than 80% of the time) give up to 2 * monshd, on
average 3/2 * monshd. In lair, when characters are often around 100 hp a
death yak heals for an average of approximately 20% health(!).

This commit halves the value of each heal
replacing hd + random2(hd) with (1 + hd) / 2 + random2(hd / 2); rounding
up always for the base so that something happens when the effect is
triggered.

[ Resolves #1878 ]