Extend berserk duration on every kill with Wrath of Trog
Intended to make the weapon more appealing for characters who can go
berserk without making it too powerful for those who can circumvent
the *Rage.  The duration extension is identical to that of the
Necklace of Bloodlust, but without the coinflip().  A flavor message
appears about 25% of the time to avoid message spam; this may have the
unintended effect of communicating that the duration extension only
occurs sometimes.