Make the Wrath of Trog angrier
This adds two new behaviors intended to make the weapon more appealing
for characters who can go berserk without making it too powerful for
those who can circumvent the *Rage. The first is berserk duration
extension identical to that of the Necklace of Bloodlust. The second is
Rampage: rampaging doesn't incur berserk penalties if it is used to take
a swing.

This is an overlap of effect, but the Wrath of Trog offers no optional
berserk and the ``extras'' are cleaner. Bloodlust might need some design
work, but that's out of scope for this commit.

To prevent using the weapon to go berserk on demand and otherwise not
deal with berserking in every fight, this also adds an on-wield penalty:
the berserk cool down.

[ Committer's note: Closes #1890. Did some commit re-ordering and
  editing. Added the wield penalty. ]