Adjust sacrifice skill.
Allow the player to take it up to three times. The mutation already supported
this, but a check in _sacrifice_is_possible prevented ru from offering it again
once it was taken. I am not sure if this was intentional or not. This commit
also adjusts the piety downward somewhat for the first sacrifice, but more
piety is given if the player elects to take it multiple times.