Add a font option for the consumables panel (nojaa3)
The pixel font used for displaying the quantities of items is quite
small and doesn't scale well. Instead of mimicking the style of
the inventory panel from local tiles, it's better to use a system font
and allow the player to configure it.

This commit adds an option to specify the font for the panel and
increases the default font size.