Remove food(.des)
food.des housed the guaranteed food vaults, which after the removal of
food simply became a part of the minivault pool.

This moves most of them to mini_monsters.des.

The spriggan baker becomes the spriggan hatter, placing hats and
occasionally a skinned spriggan corpse (yikes!).

The "fruiting plant" subvaults are removed (they had a TODO asking for
flavour without food). Vaults that use them are adjusted.

One vault that was a notorious and spoilery autoexplore trap is removed.

Another, with jelly flavour, is removed. The flavour wasn't strong
without being able to place royal jellies, and without players eating at
all the side-by-side jellies eating the door and a ration for the player
is also lost.