Add demonic trees (Sastreii)
A new type of tree appropriate for the Abyss and Pan. When ignited, it
releases a durably summoned hostile demon and leaves a clouds of chaos.
The demon has a chance to be a greater demon based on the depth of tree
placement. The chaos clouds are single tile chaos clouds with somewhat
long duration, but don't spread like forest fires. We now place these as
the default tree type in both Abyss and Pan. These trees work normally
in terms of being awoken and susceptibility to fire/lightning bolts.
They have a set of very creative tiles, courtesy of Sastreii, a
description, as well as a description of their unique burning behaviour.