Mark diamond obelisks as not a threat
The only effect of this should be allow ghouls / bloodless vp to heal
with obelisks in sight. Their effect is via clouds, which should still
prevent autotravel etc.

Also, remove some tornado dprf spam. Easy enough for someone adding
tornado to put in temporarily if needed later.

Possibly resolves mantis 12569, but not any hepl bug (such as #1747).
That is, I suspect 12569 was not actually hepl related, but that there
is another similar hep bug.