Force-interrupt all delays on zot clock messages
One factor in the death in 12569 is that when you have
`explore_auto_rest = true`, and start autotravel, it uses the delay
interrupt options for `interrupt_travel` rather than `interrupt_run`,
which does not include messages. It seems pretty reasonable that the zot
clock alerts should just interrupt anything, regardless of
configuration, which sidesteps this issue.