Small tweaks to LCS effect tiles
Small adjustments in symmetry and centering of the LCS effect animation tiles.
Specifically crystal_spear1.png was 27x27 instead of 32x32 and had a stray pixel
in lower left corner, which may have been causing a display glitch in webtiles.
Some of the other tiles were not centered properly and had varying transparency
in the different directions (no. 7 was notably shorter than the others).