fix: check for exclusions in view.is_safe_square
Currently view.is_safe_square returns false if a square has clouds or a
trap or something known to be nasty, but not if the square is marked
excluded. However, a marked exclusion, whether by the player or the
game, would seem to be an obvious indicator that the square is unsafe.

With the current implementation of view.is_safe_square, hitting TAB to
autofight when a monster is within an excluded area, such as that
surrounding an ice statue, can result in the question "Are you sure you
want to move into that travel-excluded area?" even if non-excluded
targets are available. With the advent of the travel_one_unsafe_move
option, autofights inconsistent treatment of exclusions is an interface
screw. This commit addresses that issue.

[ Committer's note: Rewrote commit message. ]